Adds support for multiple rocks
Transformed the cookiecutter template to support adding multiple rocks.

Now, the template will ask for the number of rocks, and generate the
base project files. After which, during the post_gen_project hook, it
will generate the rocks using the previous cookiecutter template.

The main downside to this approach is that rock-related extra content
cannot be injected into the template beforehand.